The only valid competition is TWWC2017
